FIX: Allow `readthedocs` to wrap tables
There is a known bug where `readthedocs` doesn't allow table cells
to wrap.  As a result, the horizontal scroll bars are very long
making it difficult to view the table.  Details on this fix
was found from:
https://rackerlabs.github.io/docs-rackspace/tools/rtd-tables.html
This anomaly is only with `readthedocs` and not seen with local
doc generation and viewing.

Fixes #108